requirements.txt
PyYAML
beautifulsoup4
requests
Jinja2
